The first line of the diff indicates that we are comparing two files, a/src/lib.rs and b/src/lib.rs. The next line shows the index values for each file; these are used by git to determine which lines have changed between the two files.
The next two lines show the actual changes made to the files. The first line, starting with ---, shows the lines that were removed from the first file. The second line, starting with +++, shows the lines that were added to the second file. In this case, a new line was added to the second file, containing the text pub mod measurement;.

The final line of the diff, starting with \ No newline at end of file, indicates that there is no newline character at the end of the second file. This is not usually considered to be a change, but git includes it in the diff to be safe.